make ladylib fit for use in submodules
