--- build.gradle
@@ -1,75 +1,0 @@
-buildscript {
-    repositories {
-        jcenter()
-        maven {
-            name = "forge"
-            url = "https://files.minecraftforge.net/maven"
-        }
-    }
-    dependencies {
-        classpath "net.minecraftforge.gradle:ForgeGradle:2.3-SNAPSHOT"
-    }
-}
-
-apply plugin: "net.minecraftforge.gradle.forge"
-
-version = modVersion
-group = modGroup
-archivesBaseName = modBaseName
-
-sourceCompatibility = 1.8
-
-sourceSets {
-    main {
-        jar {
-            exclude 'testmod/**'
-            exclude 'assets/test_mod/**'
-        }
-    }
-}
-
-minecraft {
-    version = project.forgeVersion
-    runDir = "run"
-
-    mappings = project.mcpVersion
-}
-
-repositories {
-    mavenCentral()
-    maven {
-        // location of the maven that hosts JEI files
-        name = "Progwml6 maven"
-        url = "http://dvs1.progwml6.com/files/maven"
-    }
-    maven {
-        // location of a maven mirror for JEI files, as a fallback
-        name = "ModMaven"
-        url = "modmaven.k-4u.nl"
-    }
-}
-
-dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
-    deobfCompile "mezz.jei:jei_${mcVersion}:${jeiVersion}:api"
-//    runtime "mezz.jei:jei_${mcVersion}:${jeiVersion}"
-}
-
-processResources {
-    // this will ensure that this task is redone when the versions change.
-    inputs.property "version", project.version
-    inputs.property "mcversion", project.minecraft.version
-
-    // replace stuff in mcmod.info, nothing else
-    from(sourceSets.main.resources.srcDirs) {
-        include "mcmod.info"
-
-        // replace version and mcversion
-        expand "version": project.version, "mcversion": project.minecraft.version
-    }
-
-    // copy everything else, thats not the mcmod.info
-    from(sourceSets.main.resources.srcDirs) {
-        exclude "mcmod.info"
-    }
-}
