--- gradle.properties
@@ -1,7 +1,0 @@
-modGroup=Ladysnake
-modVersion=1.0-SNAPSHOT
-modBaseName=Ladylib
-mcVersion=1.12.2
-forgeVersion=1.12.2-14.23.2.2618
-mcpVersion=snapshot_20180217
-jeiVersion=4.8.5.138
