--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Feb 17 16:57:06 CET 2018
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.5-rc-2-all.zip
