+++ ladylib/client/ICustomLocation.java
