+++ ladylib/registration/AutoRegister.java
