+++ ladylib/registration/AutoRegistrar.java
