+++ ladylib/registration/AutoRegistryRef.java
