+++ ladylib/registration/BlockRegistrar.java
