--- src/main/java/testmod/Configuration.java
@@ -1,9 +1,0 @@
-package testmod;
-
-public class Configuration {
-    public static Client client = new Client();
-
-    public static class Client {
-        public int maxParticles = 500;
-    }
-}
