--- src/main/resources/assets/ladylib/blockstates/sample_blockstate.json
@@ -1,6 +1,0 @@
-{
-  "forge_marker": 1,
-  "variants": {
-    "normal": {"model": "@NAME@"}
-  }
-}
