--- src/main/resources/assets/ladylib/models/item/sample_item.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "@DOMAIN@:items/@NAME@"
-  }
-}
