--- src/main/resources/assets/test_mod/blockstates/test_block.json
@@ -1,10 +1,0 @@
-{
-  "variants": {
-    "normal": [
-      { "model": "dirt" },
-      { "model": "dirt", "y": 90 },
-      { "model": "dirt", "y": 180 },
-      { "model": "dirt", "y": 270 }
-    ]
-  }
-}
