--- src/main/resources/assets/test_mod/models/item/hidden.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "items/redstone_dust"
-  }
-}
