--- src/main/resources/assets/test_mod/models/item/test.json
@@ -1,6 +1,0 @@
-{
-  "parent": "item/generated",
-  "textures": {
-    "layer0": "items/diamond"
-  }
-}
